Latest build 6/7/2016 11:17PM
-Fixed an error with the “rm” shell script where the “temp” metadata
file was not being removed because the previous version of the script
had it looking for “temp” to be the last word in the filename before
the extension, when in fact it is followed by a dash and a number.
This fix accommodates the dash and the number so the file is deleted
properly.  Also removed comments about this line, and entirely removed
old “Finder” method of deleting that same file, which had been
commented before now just in case it was needed again.

-Updated copyright statement in info.plist file to fix years and to
include “some rights reserved”