Latest build 6/14/2016 7:09AM
- When the curl % xferd error occurs, it now moves the offending file
to an error folder which has been added to resources, and appends a
code to the filename so I know exactly what is causing the error later.
 This will be modified in later versions, once I find out what exactly
causes the errors (and therefore how to handle them better).  Also in
future versions, the error folder will have its files processed again
(after the DownloadComplete queue is finished being processed) since
sometimes the error doesn’t occur when the file is processed at a later
time.

- Added error handling for “resource not found” error.  It is caused by
tmdb metadata variables not populating correctly for certain shows, so
certain variables were moved within a try command, and a second “do
shell script” for metadata was created specifically to handle shows
that don’t have all of their metadata variables populated properly.

- Also fixed bug where files wouldn’t delete from processing folder
after being added to iTunes if the show was no longer in the “show
list” (list of shows in UI that the user has subscribed to).