Latest build 1/24/2017 8:18AM
- Updated code to fix problems caused by the changes torrentz2.eu made
to the format of their rss feed.
- Commented section of code that checks a show’s rss feed for the
highest episode number available, since the script now just increments
the episode number until the feed returns nothing.  If everything runs
smoothly while it is commented (which it has so far), the commented
section will be removed entirely.
- Began using curl as a replacement for the feedchecker workflow in
some places, with the goal being to completely eliminate the
feedchecker workflow if possible.