Latest build 1/31/2017 3:56AM
- Added better support for shows with apostrophes in their title—now
retrieves full name, including apostrophes, of TV shows from TMDB to
use when searching for show artwork, which requires the exact show
name.  The apostrophes are eliminated from the show name earlier in the
code for the GUI dropdown and when searching for torrents, which do not
use apostrophes.
- Increased time between feed checks from 5 minutes to 10 minutes to
avoid torrentz2.eu’s “too many queries” response, which prohibits more
than 100 requests in a 10 minute period.
- Added “repeat while” with 10 second delay if torrentz2.eu returns a
“too many queries” error for more than 5 requests in the last 10
seconds; continues searching feeds when enough time has passed and it
stops returning the error.
- When checking if a .torrent file has already been downloaded for an
episode, it now only checks the show name and episode code rather than
the full .torrent filename, so multiple versions of the same episode
with different video/audio quality won’t download simultaneously.  If a
higher quality version of the same episode becomes available, it will
download later, after the first version to download has already been
added to iTunes.  This way, some version of the episode is added to
iTunes as soon as possible.