Latest build 2/1/2017 7:03PM
- Added code that moves or deletes files from the processing folder
that have failed to process and are stuck there, preventing all
subsequent completed downloads from processing.
- Episode IDs in the GUI are now incremented both when they are added
to iTunes and when they are deleted for being a duplicate of an episode
already in iTunes in the same or better quality.  In the past, the same
episode would download and delete over and over again because the
episode ID in the GUI would only be incremented when episodes were
added to iTunes, not when they were deleted for being duplicates.  -
Since the code that increments episode IDs is now used in two places
(see above), it has been made into a subroutine so it can be called
multiple times.