Latest build 2/9/2017 4:51am
- Modified code that moves or deletes “stuck” files from the processing
folder so that it now moves the oldest video file back to the
downloading complete directory, and deletes everything else, rather
than figuring out what to move for every possible combination of video
files.  Only exception is when mkv and m4v files are both there, but
there are no files in the rawfiles directory, which means the m4v was
created successfully, so it keeps that and deletes the mkv even though
the m4v is newer.
- Greatly simplified the code in the download: subroutine so that it
has much fewer unnecessary lines and variables.