Latest build 2/16/2017 4:22PM
- Includes latest version of ffmpeg
- Now uses ffmpeg alone to re-mux mkv files to m4v instead of using
mkvextract, faac, and mp4box
- Removed RawStreams directory since it is no longer needed.
- updateEpcode: is now called before trashTorrent:, and both have been
moved out of the try/on error block to ensure that the episode code in
the show list is updated before the torrent file is deleted so a
duplicate of the same episode is not added in the split second between
the deletion of the torrent file and the update of the episode code.
- Line of code that checks if a torrent file has already been
downloaded before downloading it again is now moved to download:
subroutine from the grabTorrent: subroutine so it doesn’t bother
updating unnecessary lines of code.
- Added lines of code within processing: subroutine to check if the
file has been moved to the error directory before trying to act on that
file within the processing directory
- Began to simplify and delete unnecessary variables and lines of code
from processing: subroutine.