Latest build 8/16/2017 9:17AM
- Fixed a bug caused by the stuckProcess subroutine  that caused files
to move back and forth between the Processing and DownloadComplete
directories nonstop.  StuckProcess assumes that files in the Processing
directory that haven’t been modified in the past 12 minutes are no
longer processing, and thus “stuck”, when actually, in previous
versions of Episodes, they hadn’t even had a chance to begin processing
because stuckProcess ran once every 8 seconds.  It now runs every 60
seconds, giving files enough time to start processing before they are
assumed to be “stuck”.
- Added code to the grabTorrent subroutine to check to make sure the
<title> in the rss feed actually contains the episode code it is
looking for, because sometimes the torrentz2 feed would include results
that weren’t the actual episode the feed was meant to be looking for,
like full seasons, etc.
- Added the “^” character before the name of the TV show to the RSS url
on torrentz2.eu so it will only return torrents whose names BEGIN with
the name of the TV show, instead of just containing the name of the TV
show, since these results are usually more reliable.
- Created resumeTorrent subroutine, which runs on launch to start
downloading any torrents in the TorrentAdd directory that don’t already
have corresponding video files in another directory, like Downloading
or Processing.  In the future, it can be called at regular intervals
instead of just on launch, and will also be used to either restart or
continue downloading partially downloaded files.