Latest build 9/1/17 4:47AM
- In the grabTorrent: subroutine, when going through potential feed
items, instead of just making the setting of the “theHash” variable
dependent on whether or not the name of the torrent contains the show
name, the entire block is now dependent on it.  This way the feed item
is only considered at all if its name contains the show’s name, and the
“theHash” variable never is called if it would have been blank anyway
(this threw an error in the past).