Latest build 2/17/2018 1:27AM
- Torrents that cannot be opened because they have no data in them (0
bytes) are now deleted from the torrentadd directory, and their hashes
are added to the “blacklist”, a .txt file that the grabTorrent
subroutine now checks each torrent hash against before downloading so
it doesn’t download the same bad file more than once.  This forces it
to choose a different item in the episode’s RSS feed.
- Removed title_appendage variable, which only served to concatenate
other variables.